+++ Changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### Fixed
+ - Login action returning the wrong results.
+ - Current not checking access to deny anonymous.
+ - Fixes to entity permissions roles.
+
 ## 0.4.2
 ### Fixed
  - Save as reference to save and load correctly.