+++ Changelog.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.4.3
 ### Fixed
  - Login action returning the wrong results.
  - Current not checking access to deny anonymous.